fix: split visual group layer geometry to satisfy file line limit
